requirements.txt
pip~=21.3.1
wheel~=0.37.1
cryptography~=36.0.1
lxml~=4.7.1
pdfkit~=1.0.0
urllib3~=1.26.8
six~=1.16.0
PyMsgBox~=1.0.9
selenium~=4.1.0
pyperclip~=1.8.2
Pillow~=9.0.0
MouseInfo~=0.1.3
PyScreeze~=0.1.28
pytweening~=1.0.4
PyGetWindow~=0.0.9
PyAutoGUI~=0.9.53
numpy~=1.22.1
setuptools~=60.5.0
PyRect~=0.1.4
h11~=0.12.0
cffi~=1.15.0
pytz~=2021.3
trio~=0.19.0
outcome~=1.1.0
attrs~=21.4.0
sniffio~=1.2.0
sortedcontainers~=2.4.0
idna~=3.3
pandas~=1.3.5
python-dateutil~=2.8.2
certifi~=2021.10.8
pycparser~=2.21
wsproto~=1.0.0
html5lib~=1.1
webencodings~=0.5.1
beautifulsoup4~=4.10.0
soupsieve~=2.3.1